Revert "Makefile: Fix install error with make -j option"
This reverts commit d2aae8477cd00325bb7c7c7e95be488088900c48.

It is completely and utterly broken.  Module install should not build
any files, and adding broken dependencies to "help" it build files is
complete and utter sh*t.

The kernel should not be built by root, and "make install" and "make
module_install" (that for obvious reasons need to be run as root)
absolutely must not build any files.  They should only ever copy the
already-built files over.

So having dependencies for the install targets is wrong, wrong, wrong.

If you try to install a kernel without building it first, you *should*
get errors. The build system shouldn't try to help root build the files.

Change-Id: I32363ecc1478791ed30e6764b3dbbf4462de1e33
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: placiano <placiano80@gmail.com>