sched_clock: Use an hrtimer instead of timer
In the next patch we're going to increase the number of bits that
the generic sched_clock can handle to be greater than 32. With
more than 32 bits the wraparound time can be larger than what can
fit into the units that msecs_to_jiffies takes (unsigned int).
Luckily, the wraparound is initially calculated in nanoseconds
which we can easily use with hrtimers, so switch to using an
hrtimer.

Cc: Russell King <linux@arm.linux.org.uk>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
[jstultz: Fixup hrtimer intitialization order issue]
Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Ian Maund <imaund@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Change-Id: I8b1f10e6894b89f9c4d9637752a828c09108ccdf
Signed-off-by: placiano <placiano80@gmail.com>