Provide version number for Debian firmware package
scripts/package/builddeb is used to create Debian packages.
Currently the firmware package always gets the same version number
irrespective of the Kernel version.
The paths inside the firmware package depend on the Kernel
version.

With the patch supplied the Kernel version becomes part of the
Debian firmware package number.

Change-Id: Icbd8d0a7bc8bdfd76ce9df788ac6d6ef7a64d23f
Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cc: Ben Hutchings <ben@decadent.org.uk>
Acked-by: maximilian attems <max@stro.at>
Signed-off-by: Michal Marek <mmarek@suse.cz>
Signed-off-by: placiano <placiano80@gmail.com>