qcom: mpm-of: fix bug in finding the pending interrupts
When APPS proc wakeup from RPM assisted PC states it reads
the vMPM status register to know the reason for wakeup and
retrigger the interrupts to handle it. Current code is not
caring about actual enabled interrupts in vMPM. Sometimes
this is causing APPS proc unnecessarily triggering the
interrupts which the not raised in the system. Fix this by
considering the enabled interrupts in vMPM before triggering
the interrupts.

Change-Id: I5de72560b4459e2d1eea3c9b97d4ec7e93abb059
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>