cpu: Extend cpu core control by having an option to link core4 online status to core3
mpdecision keeps core 4 online all of the time, which is a bit a waste of resources.
By implementing a new control option with value "3" for only the 4th core, the core will
go offline but its online status will be mirrored with the status of the 3rd core.
In a way, core 4 is locked to core 3 by that option.

Change-Id: I0fb78f014352d0fc7c6d55f3a8d9a499af34903d
Signed-off-by: placiano <placiano80@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>