crypto: msm: qcrypto: Close window during suspending for SMP
In the normal case, when a crypto operation is complete, driver process a
new request if there are any outstanding requests. In SMP, the suspending
may arrives in the window after driver completes the outstanding request,
and before driver starts a new operation. To close this window, a check
is added to make sure device is in BUS_HAS_BANDWIDTH to perform the next
operation.

Change-Id: I384a72e47dfa901f9c469d9db2a20d152a3d9d9b
Signed-off-by: AnilKumar Chimata <anilc@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: placiano <placiano80@gmail.com>