Add missing-syscalls wrapper:
missing-syscalls was the default wrapper that was used in a lot
older 2.x kernels.  The python gcc wrapper was recently removed
because it causes too many warnings and treats them as errors.
We really just want to compile the kernels in their current source
state without having to patch the kernel with a lot of modifications
that could potentially break future upstream changes.  If we want to write
a small kernel program the warnings will still be there for us to see
if there is a problem in the code.  We shouldn't be forced into writing
out code a specific way because linux.org says so ;)

From documentation: http://valgrind.org/docs/manual/dist.readme-missing.html

Syscall and ioctl wrappers have two jobs:

1. Tell a tool what's about to happen, before the syscall takes place.  A
   tool could perform checks beforehand, eg. if memory about to be written
   is actually writeable.  This part is useful, but not strictly
   essential.

2. Tell a tool what just happened, after a syscall takes place.  This is
   so it can update its view of the program's state, eg. that memory has
   just been written to.  This step is essential.

The "happenings" mostly involve reading/writing of memory.

Change-Id: I4498368ef1afa8a1ed99fd0d6f3a2f064a09a29a
Signed-off-by: Paul Beeler <pbeeler80@gmail.com>